---------------------------------------------------------------------- Enter Log.  Lines beginning with `CVS:' are removed automatically
Committing in .

Modified Files:
	OpenScada/ChangeLog OpenScada/configure OpenScada/configure.in
	OpenScada/src/Makefile.am OpenScada/src/Makefile.in
	OpenScada/src/main.cpp OpenScada/src/tarhives.cpp
	OpenScada/src/tarhives.h OpenScada/src/tconfig.cpp
	OpenScada/src/tconfig.h OpenScada/src/tcontr.cpp
	OpenScada/src/tcontr.h OpenScada/src/tcontroller.cpp
	OpenScada/src/tcontroller.h OpenScada/src/tcontrollers.cpp
	OpenScada/src/tcontrollers.h OpenScada/src/thd.cpp
	OpenScada/src/tkernel.cpp OpenScada/src/tmessage.cpp
	OpenScada/src/tmodschedul.cpp OpenScada/src/tparam.cpp
	OpenScada/src/tparamcontr.cpp OpenScada/src/tparamcontr.h
	OpenScada/src/tparams.cpp OpenScada/src/tparams.h
	OpenScada/src/tsequrity.cpp OpenScada/src/tsequrity.h
	OpenScada/src/tspecials.cpp OpenScada/src/tspecials.h
	OpenScada/src/tsys.cpp OpenScada/src/tsys.h
	OpenScada/src/ttiparam.cpp OpenScada/src/ttiparam.h
	OpenScada/src/ttipcontroller.cpp
	OpenScada/src/ttipcontroller.h OpenScada/src/ttransports.cpp
	OpenScada/src/ttransports.h OpenScada/src/tuis.cpp
	OpenScada/src/tuis.h OpenScada/src/tvalue.cpp
	OpenScada/src/tvalue.h
	OpenScada/src/moduls/arhiv/base/base.cpp
	OpenScada/src/moduls/bd/direct_dbf/dbf.h
	OpenScada/src/moduls/bd/direct_dbf/direct_dbf.cpp
	OpenScada/src/moduls/special/test_kernel/test_kernel.cpp
	OpenScada/src/moduls/tcontr/sys/Makefile.am
	OpenScada/src/moduls/tcontr/sys/Makefile.in
	OpenScada/src/moduls/tcontr/sys/sys.cpp
	OpenScada/src/moduls/tcontr/sys/sys.h
	OpenScada/src/moduls/tcontr/virtual1/virtual1.cpp
	OpenScada/src/moduls/tcontr/virtual1/virtual1.h
	OpenScada/src/moduls/transport/socket/socket.cpp
	OpenScada/src/moduls/transport/socket/po/oscd_socket.pot
	OpenScada/src/moduls/transport/socket/po/ru.po
	OpenScada/src/moduls/transport/socket/po/uk.po
	OpenScada/src/moduls/ui/web_cfg/web_cfg.cpp
	OpenScada/src/moduls/ui/web_cfg/po/ru.po
	OpenScada/src/moduls/ui/web_cfg/po/uk.po
	OpenScada/src/po/openscada.pot OpenScada/src/po/ru.po
	OpenScada/src/po/uk.po
Added Files:
	OpenScada/doc/mod_test_rest.sxw OpenScada/src/telem.cpp
	OpenScada/src/telem.h
	OpenScada/src/moduls/tcontr/sys/po/Makefile.am
	OpenScada/src/moduls/tcontr/sys/po/Makefile.in
	OpenScada/src/moduls/tcontr/sys/po/oscd_SysContr.pot
	OpenScada/src/moduls/tcontr/sys/po/ru.po
	OpenScada/src/moduls/tcontr/sys/po/uk.po
Removed Files:
	OpenScada/clean_tree OpenScada/src/tconfigelem.cpp
	OpenScada/src/tconfigelem.h OpenScada/src/tvalueelem.cpp
	OpenScada/src/tvalueelem.h
	OpenScada/src/moduls/bd/direct_dbf/Makefile
	OpenScada/src/moduls/bd/my_sql/Makefile
	OpenScada/src/moduls/protocol/self/Makefile
	OpenScada/src/moduls/special/test_bd/Makefile
	OpenScada/src/moduls/special/test_kernel/Makefile
	OpenScada/src/moduls/special/test_transp/Makefile
	OpenScada/src/moduls/tcontr/sys/Makefile
	OpenScada/src/moduls/transport/mail/Makefile
	OpenScada/src/moduls/ui/qt_gui/Makefile
	OpenScada/src/moduls/ui/web_dbg/Makefile
----------------------------------------------------------------------
- Admin: Delete clean_tree. For clean tree use <make distclean>;
- !!!!: Add localisation to the controller SYS;
- !!!!: TArhives: Made separe BD for the value and message arhives;
- !!!!: Merge TConfigElem and TValueElem to TElem;
- !!!!: Add TVal for TValue container;
- !!!!: Add TCfg for TConfig container;
- TContr: Made direct init the XMLNode value;
- TContr: Add value parser;
- TContr: Add write mode controll;
- Change enable and start mode manipulation:
	- TController;
    	- TParamContr;
      	- TTransportS;
        - TUIS;
- TSpecialS: Make private acces to Tvalue;
- !!!!: TSYS: Make ResAlloc and apply autofreeing mode for resources (for
- TSYS: Sygnal handling moved to constructor.
- !!!!: TValue: fullish remake;
- BD::direct_dbg: Fix long size for x86_64;
- Special::test_kernel: Add value test;
- Controller::virtual_v1: Fix value;
- Controller::sys: Add: CPU, Mem, HDDTemp, LmSensors;
- Transport::socket: update output transport;
- UI::web_cfg: Add checking of change The group element after press "Apply"


git-svn-id: svn://oscada.org/trunk/OpenScada@327 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/moduls/bd/direct_dbf/dbf.h
@@ -8,7 +8,8 @@ struct db_head
 {
     char ver;			//Версия DBF (def = 3)
     char data[3];		//Дата модификации
-    long numb_rec;		//Количество записей
+    //long numb_rec;		//Количество записей
+    int numb_rec;		//Количество записей
     short len_head;		//Длина заголовка
     short len_rec;		//Длина записи
     char res[20];
@@ -18,7 +19,8 @@ struct db_str_rec
 {
     char name[11];		//Имя поля
     char tip_fild;		//Тип поля (C - ASCII; N - номер; L - логический; M - Memo; D - Data)
-    long adr_in_mem;
+    //long adr_in_mem;
+    int adr_in_mem;
     char len_fild;		//Длина поля
     char dec_field;		//Знаков после "."
     char res[14];
